After talking with Brian, I deleted StickerView.java and added drag+drop code to DrawFragment.java. For the drag+drop case, I added Toast to check if the code is working.
