+++ app/src/main/res/values/strings.xml
@@ -7,6 +7,12 @@
     <string name="title_activity_draw">DrawActivity</string>
     <string name="button_set_line_width">Set Line Width</string>
     <string name="color_dialog_fragment">color dialog</string>
+    <string name="drag_started">Action is DragEvent.ACTION_DRAG_STARTED</string>
+    <string name="drag_entered">Action is DragEvent.ACTION_DRAG_ENTERED</string>
+    <string name="drag_exited">Action is DragEvent.ACTION_DRAG_EXITED</string>
+    <string name="drag_location">Action is DragEvent.ACTION_DRAG_LOCATION</string>
+    <string name="drag_ended">Action is DragEvent.ACTION_DRAG_ENDED</string>
+    <string name="drag_drop">ACTION_DROP event</string>
 
 <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
